+++ .travis/test-script.sh
@@ -144,13 +144,13 @@ else
 fi
 
 ##
-# Create a "customfile.txt", configure the 'drupal-asset-files' extra key,
+# Create a "customfile.txt", configure the 'asset-files' extra key,
 # run the command 'composer drupal:paranoia' and check if the file has been symlinked.
 #
-echo "${MSG_INFO} Create a \"customfile.txt\" and configure the 'drupal-asset-files' extra key to check if the file has been symlinked."
+echo "${MSG_INFO} Create a \"customfile.txt\" and configure the 'asset-files' extra key to check if the file has been symlinked."
 
 touch "$SITE_APP/customfile.txt"
-composer config extra.drupal-asset-files.should-be-simlinked customfile.txt
+composer config extra.drupal-paranoia.asset-files token_list_files; sed -i -e "s/\"token_list_files\"/\[\"customfile.txt\"\]/" composer.json
 
 # Rebuild web directory.
 composer drupal:paranoia || exit 1
@@ -163,13 +163,13 @@ else
 fi
 
 ##
-# Create a "customfile.php", configure the 'drupal-asset-files' extra key,
+# Create a "customfile.php", configure the 'asset-files' extra key,
 # run the command 'composer drupal:paranoia' and check if the file has not been symlinked.
 #
-echo "${MSG_INFO} Create a \"customfile.php\" and configure the 'drupal-asset-files' extra key to check if the file has not been symlinked."
+echo "${MSG_INFO} Create a \"customfile.php\" and configure the 'asset-files' extra key to check if the file has not been symlinked."
 
 touch "$SITE_APP/customfile.php"
-composer config extra.drupal-asset-files.should-not-be-simlinked customfile.php
+composer config extra.drupal-paranoia.asset-files token_list_files; sed -i -e "s/\"token_list_files\"/\[\"customfile.php\"\]/" composer.json
 
 # Rebuild web directory.
 composer drupal:paranoia || exit 1
@@ -179,4 +179,21 @@ if [ -L "$SITE_WEB/customfile.php" ]; then
   exit 1
 else
   echo "${MSG_OK} 'composer drupal:paranoia' command did not re-create the web directory with WRONG extra symlinks"
+fi
+
+##
+# Test "excludes" config.
+# Run the command 'composer drupal:paranoia' and check if the excluded paths were not symlinked or stubbed.
+#
+echo "${MSG_INFO} Add paths to 'excludes' config and check if they have not been stubbed or symlinked."
+composer config extra.drupal-paranoia.excludes token_list_files; sed -i -e "s/\"token_list_files\"/\[\"core\/install.php\"\,\"core\/tests\"\]/" composer.json
+
+# Rebuild web directory.
+composer drupal:paranoia || exit 1
+
+if [ -f "$SITE_WEB/core/install.php" ] || [ -d "$SITE_WEB/core/tests" ]; then
+  echo "${MSG_ERROR} 'composer drupal:paranoia' command re-created the web directory with excluded files and folders"
+  exit 1
+else
+  echo "${MSG_OK} 'composer drupal:paranoia' command did not re-create the web directory with excluded files and folders"
 fi